# openapi.yaml
openapi: 3.0.3
info:
  title: Movie Suggestion API
  description: >
    Personalized movie recommendation API built in Go using clean architecture.
    Integrates with OMDB for movie data, Neo4j for graph-based recommendations,
    PostgreSQL for authentication, and AWS SQS for async movie imports.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /api/v1/movies/{id}:
    get:
      tags: [Movies]
      summary: Get movie details
      description: >
        Returns the full details of a movie, including genres, actors, directors and plot.
        Requires `movies:read` role.
      operationId: getMovie
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MovieID"
      responses:
        "200":
          description: Full movie details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/movies/{id}/watched:
    post:
      tags: [Movies]
      summary: Record watched movie
      description: Records that the authenticated user watched a movie, with a rating and reaction. Requires `movies-watch:write` role.
      operationId: recordWatched
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MovieID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordWatchedRequest"
      responses:
        "200":
          description: Watch recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/movie-import:
    post:
      tags: [Import]
      summary: Trigger movie import
      description: >
        Triggers an async import of movies from OMDB for each search term.
        For each term, up to `maxPages` pages (10 results/page) are fetched
        and each movie is enqueued in SQS for processing.
        Requires `movies:write` role.
      operationId: triggerImport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerImportRequest"
      responses:
        "202":
          description: Import triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: import triggered
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    MovieID:
      name: id
      in: path
      required: true
      description: Movie UUID
      schema:
        type: string
        format: uuid
        example: 4e5ebac2-d48d-4fc6-9319-dd52cf3be8b6
  responses:
    BadRequest:
      description: Bad request
      content:
        text/plain:
          schema:
            type: string
            example: invalid request body
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        text/plain:
          schema:
            type: string
            example: forbidden
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema:
            type: string
            example: not found
    InternalError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: internal server error
  schemas:
...
    Genre:
      type: object
      properties:
        Name:
          type: string
          example: Action
    Actor:
      type: object
      properties:
        Name:
          type: string
          example: Keanu Reeves
    Director:
      type: object
      properties:
        Name:
          type: string
          example: Lana Wachowski
    Movie:
      type: object
      properties:
        ID:
          type: string
          format: uuid
          example: 4e5ebac2-d48d-4fc6-9319-dd52cf3be8b6
        Title:
          type: string
          example: The Matrix
        Year:
          type: string
          example: "1999"
        Plot:
          type: string
          example: A computer hacker learns from mysterious rebels about the true nature of his reality.
        Runtime:
          type: string
          example: 136 min
        Poster:
          type: string
          format: uri
          example: https://m.media-amazon.com/images/M/...
        ImdbRating:
          type: number
          format: float
          example: 8.7
        ImdbID:
          type: string
          example: tt0133093
        Genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        Actors:
          type: array
          items:
            $ref: "#/components/schemas/Actor"
        Directors:
          type: array
          items:
            $ref: "#/components/schemas/Director"
...
    RecordWatchedRequest:
      type: object
      required: [rating, reaction]
      properties:
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 10
          example: 8.5
        reaction:
          type: string
          enum: [liked, disliked, neutral]
          example: liked
    WatchedResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          example: 90b4fd08-ee46-4d54-a7c0-1bab4857cec3
        movieId:
          type: string
          format: uuid
          example: 4e5ebac2-d48d-4fc6-9319-dd52cf3be8b6
        rating:
          type: number
          format: float
          example: 8.5
        reaction:
          type: string
          example: liked
        watchedAt:
          type: string
          format: date-time
          example: "2026-05-30T14:30:00Z"
    TriggerImportRequest:
      type: object
      required: [searchTerms]
      properties:
        searchTerms:
          type: array
          items:
            type: string
          minItems: 1
          example: ["inception", "matrix", "interstellar"]
        maxPages:
          type: integer
          minimum: 1
          default: 1
          description: Number of OMDB result pages to fetch per search term (10 movies/page).
          example: 3
...